added new class FindExamples.cs (extracted tests from ApiExamples.cs)
