+++ src/Coypu.AcceptanceTests/Coypu.AcceptanceTests.csproj
@@ -74,6 +74,7 @@
   <ItemGroup>
     <Compile Include="Examples\ApiExamples.cs" />
     <Compile Include="Examples\CustomBrowserSession.cs" />
+    <Compile Include="Examples\FindExamples.cs" />
     <Compile Include="FindAllWithPredicateAndRetries.cs" />
     <Compile Include="Properties\ParallelizableAssemblyInfo.cs" />
     <Compile Include="Screenshots.cs" />
